requirements.txt
python-dotenv
snowflake-connector-python
sqlglot
click
fastapi
uvicorn
